+++ crates/cast/tests/integration/wait_for_tx.rs
@@ -3,10 +3,10 @@ use crate::helpers::{
     fixtures::create_test_provider,
 };
 use camino::Utf8PathBuf;
-use cast::helpers::constants::DEFAULT_RETRIES;
-use cast::{get_account, handle_wait_for_tx, parse_number, wait_for_tx};
+use cast::{get_account, helpers::constants::DEFAULT_RETRIES};
+use cast::{handle_wait_for_tx, parse_number, wait_for_tx};
+use starknet::contract::ContractFactory;
 use starknet::core::types::FieldElement;
-use starknet::{contract::ContractFactory, core::chain_id};
 
 #[tokio::test]
 async fn test_happy_path() {
@@ -29,8 +29,9 @@ async fn test_rejected_transaction() {
         ACCOUNT,
         &Utf8PathBuf::from(ACCOUNT_FILE_PATH),
         &provider,
-        chain_id::TESTNET,
+        &None,
     )
+    .await
     .expect("Could not get the account");
 
     let factory = ContractFactory::new(parse_number(MAP_CLASS_HASH_V1).unwrap(), account);