+++ crates/conversions/src/lib.rs
@@ -1,5 +1,6 @@
 use cairo_felt::Felt252;
 use starknet::core::types::FieldElement;
+use starknet_api::core::Nonce;
 use starknet_api::{
     core::{ClassHash, ContractAddress},
     hash::{StarkFelt, StarkHash},
@@ -9,6 +10,7 @@ pub mod class_hash;
 pub mod contract_address;
 pub mod felt252;
 pub mod field_element;
+pub mod nonce;
 pub mod short_string;
 pub mod stark_felt;
 
@@ -20,4 +22,5 @@ pub trait StarknetConversions {
     fn to_class_hash(&self) -> ClassHash;
     fn to_contract_address(&self) -> ContractAddress;
     fn to_short_string(&self) -> String;
+    fn to_nonce(&self) -> Nonce;
 }