+++ crates/forge/src/main.rs
@@ -7,14 +7,17 @@ use scarb_ui::args::PackagesFilter;
 use std::path::PathBuf;
 use tempfile::{tempdir, TempDir};
 
-use forge::run;
 use forge::{pretty_printing, RunnerConfig};
+use forge::{run, TestFileSummary};
 
 use forge::scarb::{
-    corelib_for_package, dependencies_for_package, get_contracts_map, name_for_package,
-    paths_for_package, target_name_for_package, try_get_starknet_artifacts_path,
+    config_from_scarb_for_package, corelib_for_package, dependencies_for_package,
+    get_contracts_map, name_for_package, paths_for_package, target_dir_for_package,
+    target_name_for_package, try_get_starknet_artifacts_path,
 };
+use forge::test_case_summary::TestCaseSummary;
 use std::process::{Command, Stdio};
+
 mod init;
 
 static PREDEPLOYED_CONTRACTS: Dir = include_dir!("crates/cheatnet/predeployed-contracts");
@@ -23,15 +26,14 @@ static PREDEPLOYED_CONTRACTS: Dir = include_dir!("crates/cheatnet/predeployed-co
 #[command(version)]
 struct Args {
     /// Name used to filter tests
-    test_name: Option<String>,
+    test_filter: Option<String>,
     /// Use exact matches for `test_filter`
     #[arg(short, long)]
     exact: bool,
-    /// Create a new forge project with <name> in current directory
-    #[clap(long)]
+    /// Create a new directory and forge project named <NAME>
+    #[arg(long, value_name = "NAME")]
     init: Option<String>,
-
-    /// Stop test execution after first failed test
+    /// Stop test execution after the first failed test
     #[arg(short = 'x', long)]
     exit_first: bool,
 
@@ -47,10 +49,19 @@ fn load_predeployed_contracts() -> Result<TempDir> {
     Ok(tmp_dir)
 }
 
-fn main_execution() -> Result<()> {
+fn extract_failed_tests(tests_summaries: Vec<TestFileSummary>) -> Vec<TestCaseSummary> {
+    tests_summaries
+        .into_iter()
+        .flat_map(|test_file_summary| test_file_summary.test_case_summaries)
+        .filter(|test_case_summary| matches!(test_case_summary, TestCaseSummary::Failed { .. }))
+        .collect()
+}
+
+fn main_execution() -> Result<bool> {
     let args = Args::parse();
     if let Some(project_name) = args.init {
-        return init::run(project_name.as_str());
+        init::run(project_name.as_str())?;
+        return Ok(true);
     }
 
     let predeployed_contracts_dir = load_predeployed_contracts()?;
@@ -63,31 +74,31 @@ fn main_execution() -> Result<()> {
 
     let scarb_metadata = MetadataCommand::new().inherit_stderr().exec()?;
 
-    let current_dir = std::env::current_dir().context("Failed to get current directory")?;
-
-    let build_output = Command::new("scarb")
-        .current_dir(current_dir)
-        .arg("build")
-        .stderr(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .output()
-        .context("Failed to build contracts with Scarb")?;
-    if !build_output.status.success() {
-        bail!(
-            "Scarb build didn't succeed:\n\n{}",
-            String::from_utf8_lossy(&build_output.stdout)
-        )
-    }
-
     let packages: Vec<PackageMetadata> = args
         .packages_filter
         .match_many(&scarb_metadata)
         .context("Failed to find any packages matching the specified filter")?;
 
-    for (index, package) in packages.iter().enumerate() {
-        let forge_config =
-            forge::scarb::config_from_scarb_for_package(&scarb_metadata, &package.id)?;
+    let mut all_failed_tests = vec![];
+    for package in &packages {
+        let forge_config = config_from_scarb_for_package(&scarb_metadata, &package.id)?;
         let (package_path, lib_path) = paths_for_package(&scarb_metadata, &package.id)?;
+        std::env::set_current_dir(package_path.clone())?;
+
+        let target_dir = target_dir_for_package(&package_path)?; // TODO(#671)
+
+        let current_dir = std::env::current_dir().context("Failed to get current directory")?;
+
+        let build_output = Command::new("scarb")
+            .current_dir(current_dir)
+            .arg("build")
+            .stderr(Stdio::inherit())
+            .stdout(Stdio::inherit())
+            .output()
+            .context("Failed to build contracts with Scarb")?;
+        if !build_output.status.success() {
+            bail!("Scarb build did not succeed")
+        }
 
         let subpackages: Vec<Utf8PathBuf> = scarb_metadata
             .packages
@@ -96,26 +107,23 @@ fn main_execution() -> Result<()> {
             .filter(|path| path.starts_with(&package_path) && *path != package_path)
             .collect();
 
-        std::env::set_current_dir(package_path.clone())?;
-
         let package_name = name_for_package(&scarb_metadata, &package.id)?;
         let dependencies = dependencies_for_package(&scarb_metadata, &package.id)?;
         let target_name = target_name_for_package(&scarb_metadata, &package.id)?;
         let corelib_path = corelib_for_package(&scarb_metadata, &package.id)?;
         let runner_config = RunnerConfig::new(
-            args.test_name.clone(),
+            args.test_filter.clone(),
             args.exact,
             args.exit_first,
             &forge_config,
         );
-
-        let contracts_path = try_get_starknet_artifacts_path(&package_path, &target_name)?;
+        let contracts_path = try_get_starknet_artifacts_path(&target_dir, &target_name)?;
         let contracts = contracts_path
             .map(|path| get_contracts_map(&path))
             .transpose()?
             .unwrap_or_default();
 
-        run(
+        let tests_file_summaries = run(
             &package_path,
             &package_name,
             &subpackages,
@@ -127,9 +135,8 @@ fn main_execution() -> Result<()> {
             &predeployed_contracts,
         )?;
 
-        if index < packages.len() - 1 {
-            pretty_printing::print_separator();
-        }
+        let mut failed_tests = extract_failed_tests(tests_file_summaries);
+        all_failed_tests.append(&mut failed_tests);
     }
 
     // Explicitly close the temporary directories so we can handle the errors
@@ -140,15 +147,18 @@ fn main_execution() -> Result<()> {
         )
     })?;
 
-    Ok(())
+    pretty_printing::print_failures(&all_failed_tests);
+
+    Ok(all_failed_tests.is_empty())
 }
 
 fn main() {
     match main_execution() {
-        Ok(()) => std::process::exit(0),
+        Ok(true) => std::process::exit(0),
+        Ok(false) => std::process::exit(1),
         Err(error) => {
             pretty_printing::print_error_message(&error);
-            std::process::exit(1);
+            std::process::exit(2);
         }
     };
 }