+++ crates/forge/tests/e2e/workspaces.rs
@@ -1,25 +1,23 @@
 use std::path::PathBuf;
 
-use assert_fs::fixture::PathCopy;
 use indoc::indoc;
 
-use crate::e2e::common::runner::runner;
+use crate::e2e::common::runner::{runner, setup_hello_workspace, setup_virtual_workspace};
 
 #[test]
 fn run_without_arguments() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
+    let temp = setup_hello_workspace();
 
     let snapbox = runner();
-
     snapbox
         .current_dir(&temp)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
         [..]Finished[..]
+
+
         Collected 3 test(s) and 2 test file(s) from hello_workspaces package
         Running 1 inline test(s)
         [PASS] hello_workspaces::test_simple
@@ -31,15 +29,15 @@ fn run_without_arguments() {
         
         [SKIP] test_failing::test_another_failing
         Tests: 1 passed, 1 failed, 1 skipped
+        
+        Failures:
+            test_failing::test_failing
         "#});
 }
 
 #[test]
 fn run_specific_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--package").arg("addition");
 
     snapbox
@@ -48,46 +46,58 @@ fn run_specific_package() {
         .success()
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
+        [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+
+
+        Collected 5 test(s) and 3 test file(s) from addition package
         Running 1 inline test(s)
         [PASS] addition::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
         "#});
 }
 
 #[test]
 fn run_specific_package2() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--package").arg("fibonacci");
 
     snapbox
         .current_dir(&temp)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
+        [..]Compiling[..]
         [..]Finished[..]
-        Collected 1 test(s) and 1 test file(s) from fibonacci package
-        Running 1 inline test(s)
+
+
+        Collected 4 test(s) and 1 test file(s) from fibonacci package
+        Running 4 inline test(s)
         [PASS] fibonacci::tests::it_works
-        Tests: 1 passed, 0 failed, 0 skipped
+        [PASS] fibonacci::tests::contract_test
+        [FAIL] fibonacci::tests::failing_test
+        
+        Failure data:
+            original value: [0], converted to a string: []
+        
+        [SKIP] fibonacci::tests::skipped_test
+        Tests: 2 passed, 1 failed, 1 skipped
+        
+        Failures:
+            fibonacci::tests::failing_test
         "#});
 }
+
 #[test]
 fn run_specific_package_and_name() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("simple").arg("--package").arg("addition");
 
     snapbox
@@ -96,31 +106,34 @@ fn run_specific_package_and_name() {
         .success()
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
+        [..]Compiling[..]
         [..]Finished[..]
-        Collected 1 test(s) and 3 test file(s) from addition package
+
+
+        Collected 2 test(s) and 3 test file(s) from addition package
         Running 0 inline test(s)
         Running 0 test(s) from tests/nested/test_nested.cairo
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 1 passed, 0 failed, 0 skipped
+        [PASS] test_simple::contract_test
+        Tests: 2 passed, 0 failed, 0 skipped
         "#});
 }
 
 #[test]
 fn run_specify_root_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--package").arg("hello_workspaces");
 
     snapbox
         .current_dir(&temp)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
         [..]Finished[..]
+
+
         Collected 3 test(s) and 2 test file(s) from hello_workspaces package
         Running 1 inline test(s)
         [PASS] hello_workspaces::test_simple
@@ -132,15 +145,15 @@ fn run_specify_root_package() {
         
         [SKIP] test_failing::test_another_failing
         Tests: 1 passed, 1 failed, 1 skipped
+        
+        Failures:
+            test_failing::test_failing
         "#});
 }
 
 #[test]
 fn run_inside_nested_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let package_dir = temp.join(PathBuf::from("crates/addition"));
 
     let snapbox = runner();
@@ -153,50 +166,66 @@ fn run_inside_nested_package() {
         [..]Compiling[..]
         [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+
+
+        Collected 5 test(s) and 3 test file(s) from addition package
         Running 1 inline test(s)
         [PASS] addition::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
         "#});
 }
 
 #[test]
 fn run_for_entire_workspace() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--workspace");
 
     snapbox
         .current_dir(&temp)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
+        [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+        
+        
+        Collected 5 test(s) and 3 test file(s) from addition package
         Running 1 inline test(s)
         [PASS] addition::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
-
-
-        Collected 1 test(s) and 1 test file(s) from fibonacci package
-        Running 1 inline test(s)
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
+        [..]Compiling[..]
+        [..]Compiling[..]
+        [..]Finished[..]
+        
+        
+        Collected 4 test(s) and 1 test file(s) from fibonacci package
+        Running 4 inline test(s)
         [PASS] fibonacci::tests::it_works
-        Tests: 1 passed, 0 failed, 0 skipped
-
-
+        [PASS] fibonacci::tests::contract_test
+        [FAIL] fibonacci::tests::failing_test
+        
+        Failure data:
+            original value: [0], converted to a string: []
+        
+        [SKIP] fibonacci::tests::skipped_test
+        Tests: 2 passed, 1 failed, 1 skipped
+        [..]Compiling[..]
+        [..]Finished[..]
+        
+        
         Collected 3 test(s) and 2 test file(s) from hello_workspaces package
         Running 1 inline test(s)
         [PASS] hello_workspaces::test_simple
@@ -208,44 +237,59 @@ fn run_for_entire_workspace() {
         
         [SKIP] test_failing::test_another_failing
         Tests: 1 passed, 1 failed, 1 skipped
+        
+        Failures:
+            fibonacci::tests::failing_test
+            test_failing::test_failing
         "#});
 }
 
 #[test]
 fn run_for_entire_workspace_inside_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let package_dir = temp.join(PathBuf::from("crates/fibonacci"));
 
     let snapbox = runner().arg("--workspace");
-
     snapbox
         .current_dir(package_dir)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
         [..]Compiling[..]
         [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+        
+        
+        Collected 5 test(s) and 3 test file(s) from addition package
         Running 1 inline test(s)
         [PASS] addition::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
-
-
-        Collected 1 test(s) and 1 test file(s) from fibonacci package
-        Running 1 inline test(s)
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
+        [..]Compiling[..]
+        [..]Compiling[..]
+        [..]Finished[..]
+        
+        
+        Collected 4 test(s) and 1 test file(s) from fibonacci package
+        Running 4 inline test(s)
         [PASS] fibonacci::tests::it_works
-        Tests: 1 passed, 0 failed, 0 skipped
-
-
+        [PASS] fibonacci::tests::contract_test
+        [FAIL] fibonacci::tests::failing_test
+        
+        Failure data:
+            original value: [0], converted to a string: []
+        
+        [SKIP] fibonacci::tests::skipped_test
+        Tests: 2 passed, 1 failed, 1 skipped
+        [..]Compiling[..]
+        [..]Finished[..]
+        
+        
         Collected 3 test(s) and 2 test file(s) from hello_workspaces package
         Running 1 inline test(s)
         [PASS] hello_workspaces::test_simple
@@ -257,18 +301,19 @@ fn run_for_entire_workspace_inside_package() {
         
         [SKIP] test_failing::test_another_failing
         Tests: 1 passed, 1 failed, 1 skipped
+        
+        Failures:
+            fibonacci::tests::failing_test
+            test_failing::test_failing
         "#});
 }
 
 #[test]
 fn run_for_entire_workspace_and_specific_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--workspace").arg("--package").arg("addition");
 
-    let result = snapbox.current_dir(&temp).assert().failure();
+    let result = snapbox.current_dir(&temp).assert().code(2);
 
     let stderr = String::from_utf8_lossy(&result.get_output().stderr);
 
@@ -277,13 +322,10 @@ fn run_for_entire_workspace_and_specific_package() {
 
 #[test]
 fn run_missing_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/hello_workspaces", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_hello_workspace();
     let snapbox = runner().arg("--package").arg("missing_package");
 
-    let result = snapbox.current_dir(&temp).assert().failure();
+    let result = snapbox.current_dir(&temp).assert().code(2);
 
     let stdout = String::from_utf8_lossy(&result.get_output().stdout);
 
@@ -292,45 +334,54 @@ fn run_missing_package() {
 
 #[test]
 fn virtual_workspace() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/virtual_workspace", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
+    let temp = setup_virtual_workspace();
     let snapbox = runner();
 
     snapbox
         .current_dir(&temp)
         .assert()
-        .success()
+        .code(1)
         .stdout_matches(indoc! {r#"
+        [..]Compiling[..]
+        [..]Finished[..]
+        
+        
+        Collected 4 test(s) and 1 test file(s) from fibonacci2 package
+        Running 4 inline test(s)
+        [PASS] fibonacci2::tests::it_works
+        [PASS] fibonacci2::tests::contract_test
+        [FAIL] fibonacci2::tests::failing_test
+        
+        Failure data:
+            original value: [0], converted to a string: []
+        
+        [SKIP] fibonacci2::tests::skipped_test
+        Tests: 2 passed, 1 failed, 1 skipped
         [..]Compiling[..]
         [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+
+
+        Collected 5 test(s) and 3 test file(s) from subtraction package
         Running 1 inline test(s)
-        [PASS] addition::tests::it_works
+        [PASS] subtraction::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
-
-
-        Collected 1 test(s) and 1 test file(s) from fibonacci package
-        Running 1 inline test(s)
-        [PASS] fibonacci::tests::it_works
-        Tests: 1 passed, 0 failed, 0 skipped
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
+        
+        Failures:
+            fibonacci2::tests::failing_test
         "#});
 }
 
 #[test]
 fn virtual_workspace_specify_package() {
-    let temp = assert_fs::TempDir::new().unwrap();
-    temp.copy_from("tests/data/virtual_workspace", &["**/*.cairo", "**/*.toml"])
-        .unwrap();
-
-    let snapbox = runner().arg("--package").arg("addition");
+    let temp = setup_virtual_workspace();
+    let snapbox = runner().arg("--package").arg("subtraction");
 
     snapbox
         .current_dir(&temp)
@@ -340,14 +391,17 @@ fn virtual_workspace_specify_package() {
         [..]Compiling[..]
         [..]Compiling[..]
         [..]Finished[..]
-        Collected 4 test(s) and 3 test file(s) from addition package
+
+
+        Collected 5 test(s) and 3 test file(s) from subtraction package
         Running 1 inline test(s)
-        [PASS] addition::tests::it_works
+        [PASS] subtraction::tests::it_works
         Running 2 test(s) from tests/nested/test_nested.cairo
         [PASS] test_nested::test_two
         [PASS] test_nested::test_two_and_two
-        Running 1 test(s) from tests/test_simple.cairo
+        Running 2 test(s) from tests/test_simple.cairo
         [PASS] test_simple::simple_case
-        Tests: 4 passed, 0 failed, 0 skipped
+        [PASS] test_simple::contract_test
+        Tests: 5 passed, 0 failed, 0 skipped
         "#});
 }